Created different folders for gui, unittests, and src code. Switched names of HomeScreen and LoginScreen.
